requirements.txt
ansi==0.2.0
atlassian-python-api==3.9.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorlog==5.0.1
cryptography==3.4.7
daemonize==2.5.0
deepmerge==0.3.0
Deprecated==1.2.12
dulwich==0.20.21
errbot==6.1.7
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
oauthlib==3.1.0
pycparser==2.20
Pygments==2.9.0
pygments-markdown-lexer==0.1.0.dev39
pyOpenSSL==20.0.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
slackclient==1.3.2
soupsieve==2.2.1
urllib3==1.26.4
waitress==2.0.0
WebOb==1.8.7
websocket-client==0.54.0
WebTest==2.0.35
Werkzeug==1.0.1
wrapt==1.12.1
